Quick demonstration of branching/first pass at RosActivityLifecycle.
